Created manual error handling
A var type is now reserved for errors, and any pointer with the lowest
bit set is considered an error. An error is then a pointer to a table
describing the error. This is the most expensive option for exception
handling and as such will probably not be used. It does have the least
element of surprise but requires significant overhead from the C side.

Note, exceptions have not be carried into the parser at this time.